+++ programs/MushroomClassifier.php
@@ -4,10 +4,10 @@ include dirname(__DIR__) . '/vendor/autoload.php';
 
 use Rubix\ML\Pipeline;
 use Rubix\ML\Datasets\Labeled;
-use Rubix\ML\NeuralNet\Layers\Dense;
 use Rubix\ML\Reports\AggregateReport;
 use Rubix\ML\Reports\ConfusionMatrix;
 use Rubix\ML\Reports\PredictionSpeed;
+use Rubix\ML\NeuralNet\Layers\Dropout;
 use Rubix\ML\NeuralNet\Optimizers\Adam;
 use Rubix\ML\Transformers\OneHotEncoder;
 use Rubix\ML\Reports\MulticlassBreakdown;
@@ -16,7 +16,7 @@ use Rubix\ML\CrossValidation\Metrics\MCC;
 use Rubix\ML\Transformers\ZScaleStandardizer;
 use Rubix\ML\Classifiers\MultiLayerPerceptron;
 use Rubix\ML\Transformers\SparseRandomProjector;
-use Rubix\ML\NeuralNet\ActivationFunctions\NoisyReLU;
+use Rubix\ML\NeuralNet\ActivationFunctions\LeakyReLU;
 use League\Csv\Reader;
 
 echo '╔═════════════════════════════════════════════════════╗' . "\n";
@@ -44,9 +44,9 @@ $labels = iterator_to_array($reader->fetchColumn('class'));
 $dataset = new Labeled($samples, $labels);
 
 $hidden = [
-    new Dense(20, new NoisyReLU(0.1)),
-    new Dense(20, new NoisyReLU(0.1)),
-    new Dense(20, new NoisyReLU(0.1)),
+    new Dropout(20, new LeakyReLU(0.01), 0.3),
+    new Dropout(20, new LeakyReLU(0.01), 0.3),
+    new Dropout(20, new LeakyReLU(0.01), 0.3),
 ];
 
 $estimator = new Pipeline(new MultiLayerPerceptron($hidden, 50, new Adam(0.001),