+++ src/NeuralNet/Layers/Binomial.php
@@ -205,9 +205,9 @@ class Binomial implements Output
             ->differentiate($this->z, $this->computed)
             ->hadamardProduct($errors);
 
-        $gradients = $errors->multiply($this->input->transpose());
+        $gradient = $errors->multiply($this->input->transpose());
 
-        $step = $optimizer->step($this->weights, $gradients);
+        $step = $optimizer->step($this->weights, $gradient);
 
         $this->weights->update($step);