+++ src/NeuralNet/Layers/Continuous.php
@@ -160,9 +160,9 @@ class Continuous implements Output
 
         $errors = new Matrix([$errors]);
 
-        $gradients = $errors->multiply($this->input->transpose());
+        $gradient = $errors->multiply($this->input->transpose());
 
-        $step = $optimizer->step($this->weights, $gradients);
+        $step = $optimizer->step($this->weights, $gradient);
 
         $this->weights->update($step);